refactor(location): change useGetFile location
